Nav bar + App name changed + Validations + Design
Started to personnalize the navigation bar
Changed the app name from Laravel to NSA
Handled the validation more precisely in association and activity forms
Improved the look of create association form

+++ app/Http/Controllers/ActivityController.php
@@ -3,11 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\CreateActivity;
-use Illuminate\Http\Request;
-use Illuminate\Http\UploadedFile;
 use App\Models\Activity;
 use App\Models\Image;
-use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Storage;
 
 class ActivityController extends Controller
 {
@@ -31,11 +29,9 @@ class ActivityController extends Controller
         $activity->location = $request->location;
         $activity->date = $request->date;
 
-        //TODO : potentially improve this required|image|mimes:jpg,png,jpeg,gif,svg|max:2048|dimensions:min_width=100,min_height=100,max_width=1000,max_height=1000
-
         if ($request->image) {
             if ($request->image->isValid()) {
-                $activity->image_id = $this->uploadImage($request->image);
+                $activity->image_id = Image::uploadImage($request->image);
             }
         }
 
@@ -58,12 +54,10 @@ class ActivityController extends Controller
         $activity->location = $request->location;
         $activity->date = $request->date;
 
-        if ($request->image)
-        {
-            if ($request->image->isValid())
-            {
+        if ($request->image) {
+            if ($request->image->isValid()) {
                 Image::destroyAndDelete($activity->image_id);
-                $activity->image_id = $this->uploadImage($request->image);
+                $activity->image_id = Image::uploadImage($request->image);
             }
         }
 
@@ -74,32 +68,12 @@ class ActivityController extends Controller
 
     public function destroy(Activity $activity)
     {
-        $activity->delete();
+        $image = Image::findOrFail($activity->id);
 
-        return redirect()->route('associations.index')->with('success', 'Association deleted successfully');
-    }
+        if (Storage::disk('public')->delete('/uploads/images/' . $image->storage_name)) {
+            $activity->delete();
+        }
 
-    public function uploadImage(UploadedFile $file)
-    {
-        $image = new Image();
-
-        $imageMimeType = $file->getMimeType();
-        $imageFullName = $file->getClientOriginalName();
-        $imageName = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-        $imageExtension = $file->getClientOriginalExtension();
-
-        $storingImageName = Str::slug($imageName . '_' . time()) . '.' . $imageExtension;
-
-        $path = $file->storeAs('uploads/images', $storingImageName, 'public');
-
-        $image->src = '/storage/' . $path;
-        $image->storage_name = $storingImageName;
-        $image->title = $imageFullName;
-        $image->alt = $imageName;
-        $image->mime_type = $imageMimeType;
-
-        $image->save();
-
-        return $image->id;
+        return redirect()->route('activities.index')->with('success', 'Activity deleted successfully');
     }
 }
